backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
neo4j==5.14.1
google-generativeai==0.3.1
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4